Creating register and login routes
