--- src/app/(site)/layout.tsx
@@ -1,13 +1,0 @@
-import Header from "@/components/landing-page/header";
-import React from "react";
-
-const HomePageLayout = ({ children }: { children: React.ReactNode }) => {
-  return (
-    <main>
-      <Header />
-      {children}
-    </main>
-  );
-};
-
-export default HomePageLayout;
