+++ src/lib/mysql/db.ts
@@ -1,12 +1,7 @@
 import { drizzle } from "drizzle-orm/mysql2";
 import "mysql2";
-import dotenv from "dotenv";
-// import schema from "../../../drizzle/schema";
-// import schema from "../../../drizzle/schema";
-import { migrate } from "drizzle-orm/mysql2/migrator";
 import { createConnection } from "mysql2";
-
-// if (!process.env.DB_URL) console.log(new Error("Cannot connect to db"));
+import * as schema from "./schema";
 
 const connection = createConnection({
   host: process.env.DB_HOST as string,
@@ -16,15 +11,5 @@ const connection = createConnection({
   database: process.env.DB_DATABASE as string,
 });
 
-const db = drizzle(connection);
-const migrateDb = async () => {
-  try {
-    console.log("Migration started!");
-    await migrate(db, { migrationsFolder: "drizzle" });
-    console.log("Migration finished succesfully!");
-  } catch (error) {
-    console.log(new Error(`Cannot migrate data ${error}`));
-  }
-};
-migrateDb();
+const db = drizzle(connection, { schema: schema, mode: "default" });
 export default db;
